+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## 9.3.0
 * Breaking change: `Document.fromHtml(html)` is now returns `Document` instead of `Delta`, use `DeltaX.fromHtml` to return `Delta`
+* Update old deprecated api from Flutter 3.19
 
 ## 9.2.14
 * feat: move cursor after inserting video/image